+++ src/main/java/org/example/hugmeexp/domain/mission/dto/response/MissionGroupResponse.java
@@ -1,4 +1,4 @@
-package org.example.hugmeexp.domain.mission.dto;
+package org.example.hugmeexp.domain.mission.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
-public class MissionGroupUpdateRequest {
+public class MissionGroupResponse {
     private Long id;
     private Long teacherId;
     private String name;
