+++ src/main/java/org/example/hugmeexp/domain/mission/service/MissionGroupService.java
@@ -1,8 +1,8 @@
 package org.example.hugmeexp.domain.mission.service;
 
-import org.example.hugmeexp.domain.mission.dto.MissionGroupRequest;
-import org.example.hugmeexp.domain.mission.dto.MissionGroupResponse;
-import org.example.hugmeexp.domain.mission.dto.MissionGroupUpdateRequest;
+import org.example.hugmeexp.domain.mission.dto.request.MissionGroupRequest;
+import org.example.hugmeexp.domain.mission.dto.response.MissionGroupResponse;
+import org.example.hugmeexp.domain.mission.dto.request.MissionGroupUpdateRequest;
 import java.util.List;
 
 public interface MissionGroupService {
